pyproject.toml
[project]
name = "travel-planner"
version = "0.1.0"
description = "AI-powered trip planning application"
requires-python = ">=3.9"
dependencies = [
    "crewai==0.19.0",
    "crewai-tools==0.0.12",
    "langchain>=0.1.0",
    "python-dotenv",
    "tiktoken==0.5.2",
    "regex==2023.12.25",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88 